+++ README.md
@@ -1,2 +1,24 @@
-# node.js
+# Node.js lessons
+Уроки от платформы GeekBrains по обучению Node js.
+Программа:
+
+Урок 1
+Введение в Node.js. Управление зависимостями
+
+Урок 2
+Цикл событий. События в Node.js
+
+Урок 3
+Работа с файловой системой. Класс Buffer. Модуль Streams
+
+Урок 4
+CLI-приложения
+
+Урок 5
+HTTP-cервер на Node.js
+
+Урок 6
+Библиотека Socket.io. Workers в Node.js
+
+
