+++ index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Node JS HomeWork 5</title>
+</head>
+<body>
+    <ul>
+        #links
+    </ul>
+</body>
+</html>
