+++ package.json
@@ -1,4 +1,22 @@
 {
-  "name": "nodehw",
-  "packageManager": "yarn@3.1.0"
+  "name": "nodehw1",
+  "version": "1.0.0",
+  "description": "",
+  "main": "homework4.js",
+  "packageManager": "yarn@3.1.0",
+  "scripts": {
+    "dev": "node homework4.js",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack-dev-server --mode development"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "colors": "^1.4.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.0.1",
+    "eslint-plugin-node": "^11.1.0"
+  }
 }
